Add files via upload
To look for an artist's upcoming shows, type in:
  node liri.js concert-this
There will be a new prompt asking what artist you would like to see. Respond with the Artist's name. 

If there are upcoming shows, LIRI will return them in the terminal.
If there are no upcoming shows scheduled, LIRI will return a message saying there is no info for that artist.

Refer to the picture attached for more help.